Напился бы да времени в обрез

+++ main.c
@@ -79,7 +79,7 @@ int main() {
 
   fclose(out);
 
-  for (int i = 0; i < N + 1; ++i) {
+  for (size_t i = 0; i < N + 1; ++i) {
     free(A[i]);
   }
